migrate tests and functionality to lib
